04-21-2023

+++ DSA/code_ninja/time&space_complexity/getTime.cpp
@@ -6,7 +6,7 @@
 #include "getTime.h"
 
 
-  long long getTimeInusec()
+uint64_t getTimeInusec()
 {
 	struct timeval current_time;
 	gettimeofday(&current_time, NULL);
