+++ DSA/code_ninja/time&space_complexity/getTime.h
@@ -2,6 +2,8 @@
 #ifndef __TIME_H
 #define __TIME_H
 
- long long getTimeInusec();
+#include<stdint.h>
+
+uint64_t getTimeInusec();
 
 #endif
